// src/zustand/rides/types.ts
export enum Types {
  RIDES_GET_ALL_REQUEST = 'RIDES_GET_ALL_REQUEST',
  RIDES_GET_ALL_FULFILLED = 'RIDES_GET_ALL_FULFILLED',
  RIDES_GET_ALL_REJECTED = 'RIDES_GET_ALL_REJECTED',

  RIDER_CHANGE_ADD_RIDES_FIELDS = "RIDER_CHANGE_ADD_RIDES_FIELDS",

  RIDES_ADD_REQUEST = 'RIDES_ADD_REQUEST',
  RIDES_ADD_FULFILLED = 'RIDES_ADD_FULFILLED',
  RIDES_ADD_REJECTED = 'RIDES_ADD_REJECTED',

  RIDES_CHANGE_FILTER_FIELDS = "RIDES_CHANGE_FILTER_FIELDS",

  RIDES_GET_ALL_CATEGORIES_FULFILLED = 'RIDES_GET_ALL_CATEGORIES_FULFILLED',

  RIDES_HANDLE_DIALOG = "RIDES_HANDLE_DIALOG",

  RIDES_DUPLICATE = "RIDES_DUPLICATE",

  RIDES_EXCLUDE_HANDLE_DIALOG = "RIDES_EXCLUDE_HANDLE_DIALOG",

  RIDES_REMOVE_REQUEST = 'RIDES_REMOVE_REQUEST',
  RIDES_REMOVE_FULFILLED = 'RIDES_REMOVE_FULFILLED',
  RIDES_REMOVE_REJECTED = 'RIDES_REMOVE_REJECTED',
}